Enable cross compiling extensions (#1994)
Two pieces here in order to enable my successful build of some pgrx
extensions from x86_64-linux to x86_64-freebsd:

- `cargo pgrx init --no-run` does not attempt to run the discovered
postgreses, since they could be for the target architecture or OS.
- `--target` flag to some other commands is passed along to cargo build
invocations and injected into the calculation of the build directory
path. The builds done in order to generate bindings (?) are still done
native, since the result has to run at build time.

Note that I was doing this within nix and nixpkgs, but I think it should
generalize pretty well to the normal rust cross toolchains.

Note also that I haven't actually tested this diff precisely, since I
was working with something that pinned, of all things, 0.12.0-alpha1.
The rebase was pretty straightforward but if you want to see the diff
that I know for sure works, check
[here](https://github.com/rhelmot/pgrx/tree/alpha-patched).